refactor(test): cleanse service test code
- separate DaoStubs from ChessServiceTest to fixture package
